Factor into start_wallet.
