+++ setup.py
@@ -83,7 +83,7 @@ kwargs = dict(
             "chia = src.cmds.chia:main",
             "chia-check-plots = src.cmds.check_plots:main",
             "chia-create-plots = src.cmds.create_plots:main",
-            "chia-wallet = src.wallet.websocket_server:main",
+            "chia-wallet = src.server.start_wallet:main",
             "chia_full_node = src.server.start_full_node:main",
             "chia_harvester = src.server.start_harvester:main",
             "chia_farmer = src.server.start_farmer:main",
